First Class Diagram:
Before starting with the Class diagram we had identified all the classes. This was done using the method of "Noun Identification Technique" explained in the book "UsingUML - Perdita Stevens". We can see the results of this technique in the file "noun_identification_technique.txt".

This commit introduces the first Class diagram that was produced from scratch with the method described above.
In this diagram 50 classes where introduced, that contain only attributes and inheritance discovered with this first approach.

By using the same book as a reference we proceed with building the CRC Cards Diagram and afterwards the complete Class Diagram.